+++ pkg/kinds/deprecated_test.go
@@ -36,6 +36,8 @@ func TestDeprecatedApiVersion(t *testing.T) {
 		{GroupVersionKind{Group: "extensions", Version: "v1beta1", Kind: "NetworkPolicy"}, true},
 		{GroupVersionKind{Group: "extensions", Version: "v1beta1", Kind: "PodSecurityPolicy"}, true},
 		{GroupVersionKind{Group: "extensions", Version: "v1beta1", Kind: "ReplicaSet"}, true},
+		{GroupVersionKind{Group: "rbac", Version: "v1alpha1", Kind: "ClusterRole"}, true},
+		{GroupVersionKind{Group: "rbac", Version: "v1beta1", Kind: "ClusterRole"}, true},
 	}
 	for _, tt := range tests {
 		t.Run(tt.gvk.String(), func(t *testing.T) {
@@ -69,8 +71,28 @@ func TestSuggestedApiVersion(t *testing.T) {
 			"networking/v1beta1/Ingress",
 		},
 		{
+			GroupVersionKind{Group: "extensions", Version: "v1beta1", Kind: "NetworkPolicy"},
+			"networking/v1/NetworkPolicy",
+		},
+		{
 			GroupVersionKind{Group: "extensions", Version: "v1beta1", Kind: "PodSecurityPolicy"},
 			"policy/v1beta1/PodSecurityPolicy",
+		},
+		{
+			GroupVersionKind{Group: "rbac", Version: "v1alpha1", Kind: "ClusterRole"},
+			"rbac/v1/ClusterRole",
+		},
+		{
+			GroupVersionKind{Group: "rbac", Version: "v1beta1", Kind: "ClusterRole"},
+			"rbac/v1/ClusterRole",
+		},
+		{
+			GroupVersionKind{Group: "scheduling", Version: "v1beta1", Kind: "PriorityClass"},
+			"scheduling/v1/PriorityClass",
+		},
+		{
+			GroupVersionKind{Group: "scheduling", Version: "v1alpha1", Kind: "PriorityClass"},
+			"scheduling/v1/PriorityClass",
 		},
 		// Current ApiVersions return the same version string.
 		{
@@ -102,6 +124,18 @@ func TestRemovedInVersion(t *testing.T) {
 		{"extensions/v1beta1:Ingress", args{
 			GroupVersionKind{Group: "extensions", Version: "v1beta1", Kind: "Ingress"},
 		}, &cluster.ServerVersion{Major: 1, Minor: 20}},
+		{"rbac/v1beta1:ClusterRole", args{
+			GroupVersionKind{Group: "rbac", Version: "v1beta1", Kind: "ClusterRole"},
+		}, &cluster.ServerVersion{Major: 1, Minor: 20}},
+		{"rbac/v1alpha1:ClusterRole", args{
+			GroupVersionKind{Group: "rbac", Version: "v1alpha1", Kind: "ClusterRole"},
+		}, &cluster.ServerVersion{Major: 1, Minor: 20}},
+		{"scheduling/v1beta1:PriorityClass", args{
+			GroupVersionKind{Group: "scheduling", Version: "v1beta1", Kind: "PriorityClass"},
+		}, &cluster.ServerVersion{Major: 1, Minor: 17}},
+		{"scheduling/v1alpha1:PriorityClass", args{
+			GroupVersionKind{Group: "scheduling", Version: "v1alpha1", Kind: "PriorityClass"},
+		}, &cluster.ServerVersion{Major: 1, Minor: 17}},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {