+++ sdk/nodejs/policy/v1beta1/PodSecurityPolicy.ts
@@ -94,6 +94,13 @@ import { getVersion } from "../../version";
               opts.version = getVersion();
           }
 
-          super(PodSecurityPolicy.__pulumiType, name, props, opts);
+          const _opts = pulumi.mergeOptions(opts, {
+              aliases: [
+                  { parent: opts.parent, type: "kubernetes:policy/v1beta1:PodSecurityPolicy", name: name },
+                  { parent: opts.parent, type: "kubernetes:extensions/v1beta1:PodSecurityPolicy", name: name },
+              ],
+          });
+
+          super(PodSecurityPolicy.__pulumiType, name, props, _opts);
       }
     }