+++ sdk/python/pulumi_kubernetes/extensions/v1beta1/NetworkPolicy.py
@@ -75,8 +75,14 @@ class NetworkPolicy(pulumi.CustomResource):
 
         __props__['status'] = None
 
+        parent = opts.parent if opts and opts.parent else None
+        aliases = [
+            pulumi.Alias(parent=parent, type_="kubernetes:networking/v1:NetworkPolicy", name=resource_name),
+            pulumi.Alias(parent=parent, type_="kubernetes:extensions/v1beta1:NetworkPolicy", name=resource_name),
+        ]
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(
             version=version.get_version(),
+            aliases=aliases,
         ))
 
         super(NetworkPolicy, self).__init__(