+++ sdk/python/pulumi_kubernetes/rbac/v1/Role.py
@@ -72,8 +72,15 @@ class Role(pulumi.CustomResource):
 
         __props__['status'] = None
 
+        parent = opts.parent if opts and opts.parent else None
+        aliases = [
+            pulumi.Alias(parent=parent, type_="kubernetes:rbac/v1:Role", name=resource_name),
+            pulumi.Alias(parent=parent, type_="kubernetes:rbac/v1beta1:Role", name=resource_name),
+            pulumi.Alias(parent=parent, type_="kubernetes:rbac/v1alpha1:Role", name=resource_name),
+        ]
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(
             version=version.get_version(),
+            aliases=aliases,
         ))
 
         super(Role, self).__init__(