+++ sdk/python/pulumi_kubernetes/rbac/v1beta1/ClusterRole.py
@@ -84,8 +84,15 @@ class ClusterRole(pulumi.CustomResource):
 
         __props__['status'] = None
 
+        parent = opts.parent if opts and opts.parent else None
+        aliases = [
+            pulumi.Alias(parent=parent, type_="kubernetes:rbac/v1:ClusterRole", name=resource_name),
+            pulumi.Alias(parent=parent, type_="kubernetes:rbac/v1beta1:ClusterRole", name=resource_name),
+            pulumi.Alias(parent=parent, type_="kubernetes:rbac/v1alpha1:ClusterRole", name=resource_name),
+        ]
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(
             version=version.get_version(),
+            aliases=aliases,
         ))
 
         super(ClusterRole, self).__init__(