+++ sdk/python/pulumi_kubernetes/rbac/v1beta1/ClusterRoleBinding.py
@@ -84,8 +84,15 @@ class ClusterRoleBinding(pulumi.CustomResource):
 
         __props__['status'] = None
 
+        parent = opts.parent if opts and opts.parent else None
+        aliases = [
+            pulumi.Alias(parent=parent, type_="kubernetes:rbac/v1:ClusterRoleBinding", name=resource_name),
+            pulumi.Alias(parent=parent, type_="kubernetes:rbac/v1beta1:ClusterRoleBinding", name=resource_name),
+            pulumi.Alias(parent=parent, type_="kubernetes:rbac/v1alpha1:ClusterRoleBinding", name=resource_name),
+        ]
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(
             version=version.get_version(),
+            aliases=aliases,
         ))
 
         super(ClusterRoleBinding, self).__init__(