+++ sdk/python/pulumi_kubernetes/scheduling/v1beta1/PriorityClass.py
@@ -115,8 +115,15 @@ class PriorityClass(pulumi.CustomResource):
 
         __props__['status'] = None
 
+        parent = opts.parent if opts and opts.parent else None
+        aliases = [
+            pulumi.Alias(parent=parent, type_="kubernetes:scheduling/v1:PriorityClass", name=resource_name),
+            pulumi.Alias(parent=parent, type_="kubernetes:scheduling/v1beta1:PriorityClass", name=resource_name),
+            pulumi.Alias(parent=parent, type_="kubernetes:scheduling/v1alpha1:PriorityClass", name=resource_name),
+        ]
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(
             version=version.get_version(),
+            aliases=aliases,
         ))
 
         super(PriorityClass, self).__init__(