+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Bug fixes
 
--   Fix deprecation warnings and docs. (https://github.com/pulumi/pulumi-kubernetes/pull/918).
+-   Fix deprecation warnings and docs. (https://github.com/pulumi/pulumi-kubernetes/pull/918 and https://github.com /pulumi/pulumi-kubernetes/pull/921).
 
 ## 1.4.0 (December 9, 2019)